add import type and rename files in folder type
