+++ src/components/Header/types/pages.ts
